pyproject.toml
[project]
name = "pydantic-ai-cloudflare"
version = "0.1.0"
description = "PydanticAI integration for Cloudflare's AI stack — Workers AI, Browser Run, Vectorize, D1, AI Gateway"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Vamshi", email = "vamshi694@users.noreply.github.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Pydantic",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "pydantic-ai-slim>=1.80.0",
    "httpx>=0.27",
    "openai>=1.60",
]

[project.optional-dependencies]
all = [
    "pydantic-ai-cloudflare[browser-run,vectorize,d1]",
]
browser-run = []
vectorize = []
d1 = []
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=0.24",
    "pytest-recording",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.urls]
Homepage = "https://github.com/vamshi694/pydantic-ai-cloudflare"
Repository = "https://github.com/vamshi694/pydantic-ai-cloudflare"
Issues = "https://github.com/vamshi694/pydantic-ai-cloudflare/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pydantic_ai_cloudflare"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["pydantic_ai_cloudflare"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
